refactor: separate algorithms to own classes
