+++ packages/client/src/signature/SignatureValidator.ts
@@ -1,15 +1,21 @@
 import { Lifecycle, scoped } from 'tsyringe'
 import { SignatureType, StreamMessage, StreamMessageError } from '@streamr/protocol'
-import { createSignaturePayload } from './signature'
-import { verifySignature } from '@streamr/utils'
 import { ERC1271ContractFacade } from '../contracts/ERC1271ContractFacade'
+import { ERC1271Validator } from './ERC1271Algorithms'
+import { SECP256K1Validator } from './SECP256K1Algorithms'
+
+export interface SignatureValidatorAlgorithm {
+    assertSignatureIsValid(streamMessage: StreamMessage): Promise<boolean>
+}
 
 @scoped(Lifecycle.ContainerScoped)
 export class SignatureValidator {
-    private readonly erc1271ContractFacade: ERC1271ContractFacade
+    private readonly validatorAlgorithms = new Map<SignatureType, SignatureValidatorAlgorithm>()
 
     constructor(erc1271ContractFacade: ERC1271ContractFacade) {
-        this.erc1271ContractFacade = erc1271ContractFacade
+        this.validatorAlgorithms.set(SignatureType.LEGACY_SECP256K1, new SECP256K1Validator())
+        this.validatorAlgorithms.set(SignatureType.SECP256K1, new SECP256K1Validator())
+        this.validatorAlgorithms.set(SignatureType.ERC_1271, new ERC1271Validator(erc1271ContractFacade))
     }
 
     /**
@@ -17,26 +23,13 @@ export class SignatureValidator {
      * Resolves if valid, rejects otherwise.
      */
     async assertSignatureIsValid(streamMessage: StreamMessage): Promise<void> {
-        const payload = createSignaturePayload({
-            messageId: streamMessage.messageId,
-            messageType: streamMessage.messageType,
-            content: streamMessage.content,
-            signatureType: streamMessage.signatureType,
-            encryptionType: streamMessage.encryptionType,
-            prevMsgRef: streamMessage.prevMsgRef ?? undefined,
-            newGroupKey: streamMessage.newGroupKey ?? undefined
-        })
+        const validatorAlgorithm = this.validatorAlgorithms.get(streamMessage.signatureType)
+        if (validatorAlgorithm === undefined) {
+            throw new Error(`Unsupported SignatureType: ${SignatureType}`)
+        }
         let success: boolean
         try {
-            if (streamMessage.signatureType !== SignatureType.ERC_1271) {
-                success = verifySignature(streamMessage.getPublisherId(), payload, streamMessage.signature)
-            } else {
-                success = await this.erc1271ContractFacade.isValidSignature(
-                    streamMessage.getPublisherId(),
-                    payload,
-                    streamMessage.signature
-                )
-            }
+            success = await validatorAlgorithm.assertSignatureIsValid(streamMessage)
         } catch (err) {
             throw new StreamMessageError(`An error occurred during address recovery from signature: ${err}`, streamMessage)
         }
